requirements.txt
Django==3.2.6
djangorestframework==3.12.4
psycopg2-binary==2.8.5
pytest-django==4.4.0
bcrypt==3.2.0
djangorestframework-simplejwt==4.8.0
celery==5.1.2
redis==3.5.3
requests==2.26.0
